Removed Variant, AltAllele.
