+++ src/main/scala/is/hail/variant/AltAllele.scala
@@ -1,10 +1,6 @@
 package is.hail.variant
 
-import is.hail.check._
 import is.hail.utils._
-import is.hail.variant.AltAlleleType.AltAlleleType
-import org.apache.spark.sql.Row
-import org.json4s._
 
 import scala.util.matching.Regex
 
@@ -18,24 +14,9 @@ object CopyState extends Enumeration {
   val Auto, HemiX, HemiY = Value
 }
 
-object AltAllele {
-
-  def gen(ref: String): Gen[AltAllele] =
-    for (alt <- Gen.frequency((10, genDNAString),
-      (1, Gen.const("*"))) if alt != ref)
-      yield AltAllele(ref, alt)
-
-  def gen: Gen[AltAllele] =
-    for (ref <- genDNAString;
-    alt <- genDNAString if alt != ref)
-      yield AltAllele(ref, alt)
-
-  implicit def altAlleleOrder: Ordering[AltAllele] = new Ordering[AltAllele] {
-    def compare(x: AltAllele, y: AltAllele): Int = x.compare(y)
-  }
-}
-
 object AltAlleleMethods {
+  import is.hail.variant.AltAlleleType._
+
   val alleleRegex: Regex = "^([ACGT]+)|\\*$".r
 
   def validate(allele: String) {
@@ -132,89 +113,4 @@ object AltAlleleMethods {
     }
     i
   }
-}
-
-trait IAltAllele {
-  def ref(): String
-
-  def alt(): String
-
-  import AltAlleleType._
-
-  def reify(): AltAllele =
-    AltAllele(ref, alt)
-
-  def altAlleleType: AltAlleleType = {
-    if (isSNP)
-      SNP
-    else if (isInsertion)
-      Insertion
-    else if (isDeletion)
-      Deletion
-    else if (isStar)
-      Star
-    else if (ref.length == alt.length)
-      MNP
-    else
-      Complex
-  }
-
-  def isStar: Boolean = alt == "*"
-
-  def isSNP: Boolean = !isStar && ((ref.length == 1 && alt.length == 1) ||
-    (ref.length == alt.length && nMismatch == 1))
-
-  def isMNP: Boolean = ref.length > 1 &&
-    ref.length == alt.length &&
-    nMismatch > 1
-
-  def isInsertion: Boolean = ref.length < alt.length && ref()(0) == alt()(0) && alt.endsWith(ref.substring(1))
-
-  def isDeletion: Boolean = alt.length < ref.length && ref()(0) == alt()(0) && ref.endsWith(alt.substring(1))
-
-  def isIndel: Boolean = isInsertion || isDeletion
-
-  def isComplex: Boolean = ref.length != alt.length && !isInsertion && !isDeletion && !isStar
-
-  def isTransition: Boolean = isSNP && {
-    val (refChar, altChar) = strippedSNP
-    (refChar == 'A' && altChar == 'G') || (refChar == 'G' && altChar == 'A') ||
-      (refChar == 'C' && altChar == 'T') || (refChar == 'T' && altChar == 'C')
-  }
-
-  def isTransversion: Boolean = isSNP && !isTransition
-
-  def nMismatch: Int = {
-    require(ref.length == alt.length, s"invalid nMismatch call on ref `${ ref }' and alt `${ alt }'")
-    (ref, alt).zipped.map((a, b) => if (a == b) 0 else 1).sum
-  }
-
-  def strippedSNP: (Char, Char) = {
-    require(isSNP, "called strippedSNP on non-SNP")
-    (ref, alt).zipped.dropWhile { case (a, b) => a == b }.head
-  }
-
-  def toRow: Row = Row(ref, alt)
-
-  def toJSON: JValue = JObject(
-    ("ref", JString(ref)),
-    ("alt", JString(alt)))
-
-  override def toString: String = s"$ref/$alt"
-
-  def compare(that: AltAllele): Int = {
-    val c = ref.compare(that.ref)
-    if (c != 0)
-      return c
-
-    alt.compare(that.alt)
-  }
-}
-
-case class AltAllele(ref: String, alt: String) extends IAltAllele {
-  require(ref != alt, "ref was equal to alt")
-  require(!ref.isEmpty, "ref was an empty string")
-  require(!alt.isEmpty, "alt was an empty string")
-
-  override def reify(): AltAllele = this
 }
