+++ src/main/scala/is/hail/variant/Variant.scala
@@ -1,66 +1,23 @@
 package is.hail.variant
 
-import is.hail.annotations.{Annotation, Region, RegionValue}
-import is.hail.check.{Arbitrary, Gen}
-import is.hail.expr.types._
+import is.hail.annotations.Annotation
+import is.hail.check.Gen
 import is.hail.utils._
-import org.apache.spark.SparkContext
-import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.Row
-import org.apache.spark.sql.types._
-import org.json4s._
-
-import scala.collection.JavaConverters._
 
 object Contig {
   def gen(rg: ReferenceGenome): Gen[(String, Int)] = Gen.oneOfSeq(rg.lengths.toSeq)
 }
 
-object Variant {
-  def apply(contig: String,
-    start: Int,
-    ref: String,
-    alt: String): Variant = {
-    Variant(contig, start, ref, Array(AltAllele(ref, alt)))
-  }
+object VariantMethods {
 
-  def apply(contig: String,
-    start: Int,
-    ref: String,
-    alt: String,
-    rg: RGBase): Variant = {
-    rg.checkVariant(contig, start, ref, alt)
-    Variant(contig, start, ref, alt)
-  }
-
-  def apply(contig: String,
-    start: Int,
-    ref: String,
-    alts: Array[String]): Variant = Variant(contig, start, ref, alts.map(alt => AltAllele(ref, alt)))
-
-  def apply(contig: String,
-    start: Int,
-    ref: String,
-    alts: Array[String],
-    rg: RGBase): Variant = {
-    rg.checkVariant(contig, start, ref, alts)
-    Variant(contig, start, ref, alts)
-  }
-
-  def apply(contig: String,
-    start: Int,
-    ref: String,
-    alts: java.util.ArrayList[String],
-    rg: RGBase): Variant = Variant(contig, start, ref, alts.asScala.toArray, rg)
-
-  def parse(str: String, rg: RGBase): Variant = {
+  def parse(str: String, rg: RGBase): (Locus, IndexedSeq[String]) = {
     val elts = str.split(":")
     val size = elts.length
     if (size < 4)
       fatal(s"Invalid string for Variant. Expecting contig:pos:ref:alt1,alt2 -- found `$str'.")
 
     val contig = elts.take(size - 3).mkString(":")
-    Variant(contig, elts(size - 3).toInt, elts(size - 2), elts(size - 1).split(","), rg)
+    (Locus(contig, elts(size - 3).toInt, rg), elts(size - 2) +: elts(size - 1).split(","))
   }
 
   def variantID(contig: String, start: Int, alleles: IndexedSeq[String]): String = {
@@ -73,18 +30,48 @@ object Variant {
     nAlleles * (nAlleles + 1) / 2
   }
 
-  def fromLocusAlleles(a: Annotation): Variant = {
-    val r = a.asInstanceOf[Row]
-    val l = r.getAs[Locus](0)
-    val alleles = r.getAs[IndexedSeq[String]](1)
-    if (l == null || alleles == null)
-      null
-    else
-      Variant(l.contig, l.position, alleles(0), alleles.tail.map(x => AltAllele(alleles(0), x)))
-  }
-
   def locusAllelesToString(locus: Locus, alleles: IndexedSeq[String]): String =
     s"$locus:${ alleles(0) }:${ alleles.tail.mkString(",") }"
+
+  def minRep(locus: Locus, alleles: IndexedSeq[String]): (Locus, IndexedSeq[String]) = {
+    val ref = alleles(0)
+
+    val altAlleles = alleles.tail
+
+    if (ref.length == 1)
+      (locus, alleles)
+    else if (altAlleles.forall(a => AltAlleleMethods.isStar(ref, a)))
+      (locus, ref.substring(0, 1) +: altAlleles)
+    else {
+      val alts = altAlleles.filter(a => !AltAlleleMethods.isStar(ref, a))
+      require(alts.forall(ref != _))
+
+      val min_length = math.min(ref.length, alts.map(x => x.length).min)
+      var ne = 0
+
+      while (ne < min_length - 1
+        && alts.forall(x => ref(ref.length - ne - 1) == x(x.length - ne - 1))
+      ) {
+        ne += 1
+      }
+
+      var ns = 0
+      while (ns < min_length - ne - 1
+        && alts.forall(x => ref(ns) == x(ns))
+      ) {
+        ns += 1
+      }
+
+      if (ne + ns == 0)
+        (locus, alleles)
+      else {
+        assert(ns < ref.length - ne && alts.forall(x => ns < x.length - ne))
+        (Locus(locus.contig, locus.position + ns),
+          ref.substring(ns, ref.length - ne) +:
+          altAlleles.map(a => if (AltAlleleMethods.isStar(ref, a)) a else a.substring(ns, a.length - ne)).toArray)
+      }
+    }
+  }
 }
 
 object VariantSubgen {
@@ -128,153 +115,4 @@ case class VariantSubgen(
         .filter(!_.contains(ref))
     } yield
       Annotation(Locus(contig, start), (ref +: altAlleles).toFastIndexedSeq)
-}
-
-trait IVariant { self =>
-  def contig(): String
-
-  def start(): Int
-
-  def ref(): String
-
-  def altAlleles(): IndexedSeq[AltAllele]
-
-  def copy(contig: String = contig, start: Int = start, ref: String = ref, altAlleles: IndexedSeq[AltAllele] = altAlleles): Variant =
-    Variant(contig, start, ref, altAlleles)
-
-  def nAltAlleles: Int = altAlleles.length
-
-  def isBiallelic: Boolean = nAltAlleles == 1
-
-  // FIXME altAllele, alt to be deprecated
-  def altAllele: AltAllele = {
-    require(isBiallelic, "called altAllele on a non-biallelic variant")
-    altAlleles()(0)
-  }
-
-  def alt: String = altAllele.alt
-
-  def nAlleles: Int = 1 + nAltAlleles
-
-  def allele(i: Int): String = if (i == 0)
-    ref
-  else
-    altAlleles()(i - 1).alt
-
-  def nGenotypes = Variant.nGenotypes(nAlleles)
-
-  def locus: Locus = Locus(contig, start)
-
-  def isAutosomalOrPseudoAutosomal(rg: RGBase): Boolean = isAutosomal(rg) || inXPar(rg) || inYPar(rg)
-
-  def isAutosomal(rg: RGBase): Boolean = !(inX(rg) || inY(rg) || isMitochondrial(rg))
-
-  def isMitochondrial(rg: RGBase): Boolean = rg.isMitochondrial(contig)
-
-  def inXPar(rg: RGBase): Boolean = rg.inXPar(locus)
-
-  def inYPar(rg: RGBase): Boolean = rg.inYPar(locus)
-
-  def inXNonPar(rg: RGBase): Boolean = inX(rg) && !inXPar(rg)
-
-  def inYNonPar(rg: RGBase): Boolean = inY(rg) && !inYPar(rg)
-
-  private def inX(rg: RGBase): Boolean = rg.inX(contig)
-
-  private def inY(rg: RGBase): Boolean = rg.inY(contig)
-
-  import CopyState._
-
-  def copyState(sex: Sex.Sex, rg: ReferenceGenome): CopyState =
-    if (sex == Sex.Male)
-      if (inXNonPar(rg))
-        HemiX
-      else if (inYNonPar(rg))
-        HemiY
-      else
-        Auto
-    else
-      Auto
-
-  def compare(that: Variant, rg: ReferenceGenome): Int = rg.compare(this, that)
-
-  def minRep: IVariant = {
-    if (ref.length == 1)
-      self
-    else if (altAlleles.forall(a => a.isStar))
-      Variant(contig, start, ref.substring(0, 1), altAlleles.map(_.alt).toArray)
-    else {
-      val alts = altAlleles.filter(!_.isStar).map(a => a.alt)
-      require(alts.forall(ref != _))
-
-      val min_length = math.min(ref.length, alts.map(x => x.length).min)
-      var ne = 0
-
-      while (ne < min_length - 1
-        && alts.forall(x => ref()(ref.length - ne - 1) == x(x.length - ne - 1))
-      ) {
-        ne += 1
-      }
-
-      var ns = 0
-      while (ns < min_length - ne - 1
-        && alts.forall(x => ref()(ns) == x(ns))
-      ) {
-        ns += 1
-      }
-
-      if (ne + ns == 0)
-        self
-      else {
-        assert(ns < ref.length - ne && alts.forall(x => ns < x.length - ne))
-        Variant(contig, start + ns, ref.substring(ns, ref.length - ne),
-          altAlleles.map(a => if (a.isStar) a.alt else a.alt.substring(ns, a.alt.length - ne)).toArray)
-      }
-    }
-  }
-
-  override def toString: String =
-    s"$contig:$start:$ref:${ altAlleles.map(_.alt).mkString(",") }"
-
-  def toRow = {
-    Row.fromSeq(Array(
-      contig,
-      start,
-      ref,
-      altAlleles.map { a => Row.fromSeq(Array(a.ref, a.alt)) }))
-  }
-
-  def toJSON: JValue = JObject(
-    ("contig", JString(contig)),
-    ("start", JInt(start)),
-    ("ref", JString(ref)),
-    ("altAlleles", JArray(altAlleles.map(_.toJSON).toList))
-  )
-}
-
-case class Variant(contig: String,
-  start: Int,
-  ref: String,
-  override val altAlleles: IndexedSeq[AltAllele]) extends IVariant {
-  require(altAlleles.forall(_.ref == ref))
-
-  /* The position is 1-based. Telomeres are indicated by using positions 0 or N+1, where N is the length of the
-       corresponding chromosome or contig. See the VCF spec, v4.2, section 1.4.1. */
-  require(start >= 0, s"invalid variant: negative position: `${ this.toString }'")
-  require(!ref.isEmpty, s"invalid variant: empty contig: `${ this.toString }'")
-
-  override def minRep: Variant = super.minRep.asInstanceOf[Variant]
-
-  def toLocusAlleles: Row = Row(locus, IndexedSeq(ref) ++ altAlleles.map(_.alt))
-
-  def alleles: IndexedSeq[String] = {
-    val a = new Array[String](nAlleles)
-    a(0) = ref
-    var i = 1
-    while (i < a.length) {
-      a(i) = altAlleles(i - 1).alt
-      i += 1
-    }
-    a
-  }
 }
