+++ src/test/scala/is/hail/annotations/AnnotationsSuite.scala
@@ -4,7 +4,6 @@ import is.hail.SparkSuite
 import is.hail.expr.types._
 import is.hail.utils._
 import is.hail.testUtils._
-import is.hail.variant.Variant
 import org.testng.annotations.Test
 
 import scala.language.implicitConversions
