+++ src/test/scala/is/hail/methods/AggregatorSuite.scala
@@ -6,7 +6,7 @@ import is.hail.expr.types._
 import is.hail.table.Table
 import is.hail.utils._
 import is.hail.testUtils._
-import is.hail.variant.{MatrixTable, VSMSubgen, Variant}
+import is.hail.variant.{MatrixTable, VSMSubgen}
 import is.hail.{SparkSuite, TestUtils}
 import org.apache.commons.math3.random.RandomDataGenerator
 import org.apache.spark.sql.Row
