+++ src/test/scala/is/hail/methods/LinearMixedRegressionSuite.scala
@@ -8,7 +8,7 @@ import is.hail.expr.types._
 import is.hail.stats._
 import is.hail.utils._
 import is.hail.testUtils._
-import is.hail.variant.{MatrixTable, Variant}
+import is.hail.variant.MatrixTable
 import is.hail.SparkSuite
 import is.hail.TestUtils
 import org.testng.annotations.Test
