+++ src/test/scala/is/hail/methods/LogisticRegressionSuite.scala
@@ -6,7 +6,6 @@ import is.hail.expr.types._
 import is.hail.table.Table
 import is.hail.utils._
 import is.hail.testUtils._
-import is.hail.variant.Variant
 import org.testng.annotations.Test
 
 class LogisticRegressionSuite extends SparkSuite {
