+++ src/test/scala/is/hail/stats/BaldingNicholsModelSuite.scala
@@ -2,7 +2,7 @@ package is.hail.stats
 
 import breeze.stats._
 import is.hail.SparkSuite
-import is.hail.variant.{Call, Variant}
+import is.hail.variant.Call
 import is.hail.testUtils._
 import org.apache.spark.sql.Row
 import org.testng.Assert.assertEquals
