+++ src/test/scala/is/hail/variant/GenotypeSuite.scala
@@ -3,6 +3,7 @@ package is.hail.variant
 import is.hail.TestUtils
 import is.hail.check.Prop._
 import is.hail.check.Gen
+import is.hail.testUtils.Variant
 import is.hail.utils._
 import org.scalatest.testng.TestNGSuite
 import org.testng.annotations.Test
